SHIT LOADS OF CHANGES
set up the new map to work with my nation and world map classes. the individual classes for each country are not needed anymore, but i haven't taken them out yet
added initialization of stat dictionaries in world map
added the initial stats of all the countries to them durring world map instantiation.
added the ability to attack a nation..... I have not implemented the stat changes but it will return a debug log if your attack is successful